file store helper creates files for each key now and hashes them

+++ m2/ScalableStorageService/src/app_kvServer/FileStoreHelper.java
@@ -4,6 +4,7 @@ import org.apache.log4j.Logger;
 
 import java.io.*;
 import java.util.concurrent.locks.ReentrantLock;
+import app_kvEcs.md5;
 
 /**
  * Created by Zabeeh on 1/26/2017.
@@ -15,7 +16,7 @@ import java.util.concurrent.locks.ReentrantLock;
  */
 public class FileStoreHelper {
 
-    private String file;
+    private String fileLocation;
     private boolean log;
     private static Logger logger = Logger.getRootLogger();
 
@@ -36,8 +37,8 @@ public class FileStoreHelper {
     /**
      * Constructor just takes a fileName and initializes a lock for all file access
      * */
-    public FileStoreHelper(String fileName, boolean log){
-        this.file = fileName;
+    public FileStoreHelper(String fileLocation, boolean log){
+        this.fileLocation = fileLocation;
         this.log = log;
         originalFileLock = new ReentrantLock();
     }
@@ -50,14 +51,10 @@ public class FileStoreHelper {
     public String FindFromFile(String key) throws Exception{
         try {
 
-            if(log){
-                logger.info("Trying to get current directory");
-            }
-
-            String currPath = System.getProperty("user.dir");
+            String fileName = md5.HashS(key);
 
             if(log){
-                logger.info("Current directory: " + currPath);
+                logger.info("File location: " + fileLocation);
             }
 
             if(log){
@@ -68,22 +65,23 @@ public class FileStoreHelper {
             if(log){
                 logger.info("Opening streams");
             }
-            FileInputStream in = new FileInputStream(file);
-            BufferedReader reader = new BufferedReader(new InputStreamReader(in));
 
-            String line = null;
+            File file = new File(fileLocation+"\\"+fileName);
 
-            // read each line from the file
-            if(log){
-                logger.info("Opened stream, beginning reading line by line");
-            }
-            while((line = reader.readLine()) != null){
-                // if line is not empty parse the xml line
+            if(!file.exists()){
+                return null;
+            } else {
+                FileInputStream in = new FileInputStream(file);
+                BufferedReader reader = new BufferedReader(new InputStreamReader(in));
+
+                String line = null;
+
                 // read each line from the file
-                if(log){
-                    logger.info("line: "+ line);
+                if (log) {
+                    logger.info("Opened stream, beginning reading line by line");
                 }
-                if(!line.isEmpty()) {
+                line = reader.readLine();
+                if (!line.isEmpty()) {
                     int keyIndex = line.indexOf("key=\"");
                     keyIndex += "key\"".length() + 1;
                     String currKey = "";
@@ -97,13 +95,13 @@ public class FileStoreHelper {
                         keyIndex++;
                     } while (currChar != '\"');
 
-                    if(log){
-                        logger.info("key from file: "+ currKey);
+                    if (log) {
+                        logger.info("key from file: " + currKey);
                     }
                     // compare the key from the line to what was passed in
                     if (currKey.equals(key)) {
 
-                        if(log){
+                        if (log) {
                             logger.info("Found the key constructing the value.");
                         }
                         // get the value of the xml entry
@@ -116,7 +114,7 @@ public class FileStoreHelper {
 
                         //if key matched return it otherwise continue reading the file
 
-                        if(log){
+                        if (log) {
                             logger.info("Returning the value.");
                         }
                         reader.close();
@@ -124,20 +122,10 @@ public class FileStoreHelper {
                         System.gc(); //FORCES the file to be closed
                         return line.substring(to_start, keyIndex);
                     }
-                } else {
-                    break;
+                    throw new Exception("Value was invalid.");
                 }
+                throw new Exception("Value was invalid.");
             }
-
-            if(log){
-                logger.info("Could not find the key returning null.");
-            }
-
-            reader.close();
-            in.close();
-            System.gc(); //FORCES the file to be closed
-
-            return null;
         } catch (Exception ex) {
             if(log){
                 logger.info(ex.getMessage());
@@ -157,7 +145,12 @@ public class FileStoreHelper {
      * */
     public FileStoreStatusType PutInFile(String key, String value) throws Exception{
         try {
-            String currPath = System.getProperty("user.dir");
+
+            String fileName = md5.HashS(key);
+
+            if(log){
+                logger.info("Current directory: " + fileLocation);
+            }
 
             if(log){
                 logger.info("Acquiring lock - Put");
@@ -168,6 +161,14 @@ public class FileStoreHelper {
             if(log){
                 logger.info("Opening stream for writing");
             }
+
+            File file = new File(fileLocation+"\\"+fileName);
+
+            if(file.exists()){
+                throw new Exception("File already exists, do upsert not put");
+            }
+
+            file.createNewFile();
 
             BufferedWriter writer = new BufferedWriter(new FileWriter(file, true));
 
@@ -181,9 +182,6 @@ public class FileStoreHelper {
             writer.flush();
             writer.close();
             System.gc();
-
-            return FileStoreStatusType.PUT_SUCCESS;
-
         } catch (Exception ex) {
             if(log){
                 logger.info(ex.getMessage());
@@ -193,6 +191,7 @@ public class FileStoreHelper {
         } finally {
             originalFileLock.unlock();
         }
+        return FileStoreStatusType.PUT_SUCCESS;
     }
 
     /**
@@ -203,99 +202,24 @@ public class FileStoreHelper {
     public FileStoreStatusType DeleteInFile(String key) throws Exception{
         try {
 
+            String fileName = md5.HashS(key);
+
+            if(log){
+                logger.info("Current directory: " + fileLocation);
+            }
+
             if(log){
                 logger.info("Acquiring lock - Delete");
             }
 
             originalFileLock.lock();
 
-            if(log){
-                logger.info("Opening streams");
-            }
+            File file = new File(fileLocation+"\\"+fileName);
 
-            FileInputStream in = null;
-            BufferedReader reader = null;
-            PrintWriter writer = null;
-            boolean success = false; //only successful if the key was found
-
-
-            try {
-
-                in = new FileInputStream(file);
-                reader = new BufferedReader(new InputStreamReader(in));
-
-                writer = new PrintWriter("./newFile.txt", "UTF-8");
-
-                if(log){
-                    logger.info("Opened streams");
-                }
-
-                String line = null;
-
-                if(log){
-                    logger.info("Opened stream, beginning reading line by line");
-                }
-
-                while ((line = reader.readLine()) != null) {
-                    if(log){
-                        logger.info("line: "+ line);
-                    }
-                    if (!line.isEmpty()) {
-                        int keyIndex = line.indexOf("key=\"");
-                        keyIndex += "key\"".length() + 1;
-                        String currKey = "";
-                        char currChar = 0;
-                        do {
-                            currChar = line.charAt(keyIndex);
-                            if (currChar != '\"') {
-                                currKey += currChar;
-                            }
-                            keyIndex++;
-                        } while (currChar != '\"');
-
-                        if(log){
-                            logger.info("key from file: "+ currKey);
-                        }
-
-                        if (currKey.equals(key) == false) {
-                            if(log){
-                                logger.info("Not the key to delete, so copy it");
-                            }
-                            // get the value of the xml entry
-                            writer.println(line);
-                            writer.flush();
-                        } else {
-                            if(log){
-                                logger.info("Found the key so, don't copy.");
-                            }
-                            success = true;
-                        }
-                    } else {
-                        break;
-                    }
-                }
-            } catch (Exception ex)
-            {
-                logger.info(ex.getMessage());
-            } finally {
-                if(reader != null) { reader.close(); }
-                if(writer != null) { writer.close();}
-                if(in != null) { in.close(); }
-                System.gc();
-            }
-
-            File old = new File("./TestFile.txt");
-            File n = new File("./newFile.txt");
-
-            old.delete();
-
-            File another = new File("./TestFile.txt");
-            n.renameTo(another);
-
-            if(success) {
-                return FileStoreStatusType.DELETE_SUCCESS;
+            if(file.exists()){
+                file.delete();
             } else {
-                return FileStoreStatusType.DELETE_FAIL;
+                throw new Exception("File does not exist");
             }
         } catch (Exception ex) {
             if(log){
@@ -306,6 +230,7 @@ public class FileStoreHelper {
         } finally {
             originalFileLock.unlock();
         }
+        return FileStoreStatusType.DELETE_SUCCESS;
     }
 
     /**
@@ -315,6 +240,11 @@ public class FileStoreHelper {
      * */
     public FileStoreStatusType UpsertInFile(String key, String value) throws Exception{
         try {
+            String fileName = md5.HashS(key);
+
+            if(log){
+                logger.info("Current directory: " + fileLocation);
+            }
 
             if(log){
                 logger.info("Acquiring lock - Update");
@@ -326,86 +256,19 @@ public class FileStoreHelper {
                 logger.info("Opening streams");
             }
 
-            FileInputStream in = null;
-            BufferedReader reader = null;
-            PrintWriter writer = null;
-            boolean success = false;
+            File file = new File(fileLocation+"\\"+fileName);
 
-            try{
-                in = new FileInputStream(file);
-                reader = new BufferedReader(new InputStreamReader(in));
-                writer = new PrintWriter("./newFile.txt","UTF-8");
-
-                String newline = createFilEntryFromKVP(key, value);
-                String line = null; //only successful if they key was found
-
-                if(log){
-                    logger.info("Opened all streams");
-                }
-
-                while((line = reader.readLine()) != null){
-                    if(log){
-                        logger.info("line: " + line);
-                    }
-                    if(!line.isEmpty()) {
-                        int keyIndex = line.indexOf("key=\"");
-                        keyIndex += "key\"".length() + 1;
-                        String currKey = "";
-                        char currChar = 0;
-                        do {
-                            currChar = line.charAt(keyIndex);
-                            if (currChar != '\"') {
-                                currKey += currChar;
-                            }
-                            keyIndex++;
-                        } while (currChar != '\"');
-
-                        if(log){
-                            logger.info("key from file: " + currKey);
-                        }
-
-                        if (currKey.equals(key) == false) {
-                            if(log){
-                                logger.info("This is not the key to update, so just copy it to new file. Key to update is " + key);
-                            }
-                            // get the value of the xml entry
-                            writer.println(line);
-                            writer.flush();
-                        } else {
-                            if(log){
-                                logger.info("This is the key to update, so update and copy it to new file.");
-                            }
-                            success = true;
-                            writer.println(newline);
-                            writer.flush();
-                        }
-                    } else {
-                        break;
-                    }
-                }
-            }  catch (Exception ex)
-            {
-                logger.info(ex.getMessage());
-            } finally {
-                if(reader != null) { reader.close(); }
-                if(writer != null) { writer.close();}
-                if(in != null) { in.close(); }
-                System.gc();
+            if(!file.exists()){
+                throw new Exception("File does not exist");
             }
 
-            File old = new File("./TestFile.txt");
-            File n = new File("./newFile.txt");
+            BufferedWriter writer = new BufferedWriter(new FileWriter(file, false));
 
-            old.delete();
+            String line = createFilEntryFromKVP(key, value);
 
-            File another = new File("./TestFile.txt");
-            n.renameTo(another);
-
-            if(success) {
-                return FileStoreStatusType.UPSERT_SUCCESS;
-            }else {
-                return FileStoreStatusType.UPSERT_FAIL;
-            }
+            writer.write(line);
+            writer.flush();
+            writer.close();
         } catch (Exception ex) {
             if(log){
                 logger.info(ex.getMessage());
@@ -415,6 +278,7 @@ public class FileStoreHelper {
         } finally {
             originalFileLock.unlock();
         }
+        return FileStoreStatusType.UPSERT_SUCCESS;
     }
 
     private String createFilEntryFromKVP(String key, String value){
